Fixear build: mover assertGoogleGroupsConfig a instrumentation.ts
El build de Next 14 fallaba con 'Can't resolve https/net/http' al
bundlear instrumentation.ts. Next lo compila para Node y Edge
runtime; aunque el guard NEXT_RUNTIME !== "nodejs" evita que se
ejecute en Edge, webpack analiza los imports estáticamente y el
import dinámico de @/lib/googleGroups igual trae googleapis/gaxios,
que usa módulos Node-only.

Fix: inline la validación en instrumentation.ts. Ahora ese archivo no
importa nada que toque googleapis. Moví los tests a
instrumentation.test.ts y saqué el export ya innecesario de
googleGroups.ts.
